Complete `bootstrap` and `render`
There are now probably-valid implementations for the core functions! The
major problem for now is that there are some massive space-leaks to do
with cancelling event handlers on deleted nodes. However, I have a plan
for this (probably just adding a "cancel" key to the records), so I'll
get that sorted and then try to get an example together.