Complete the initial draft of the watcher algebra
Untested, but it builds: there is now a handler for each case of the
watcher algebra. I think I'll probably end up adding something like
`ArraySwap Int Int` to swap two elements round, as this will make
sorting substantially easier and more performant, but this is more than
enough to prove a concept!